requirements.txt
XlsxWriter~=3.2.0
aiofiles~=0.7.0
aiohttp~=3.9.5
argparse~=1.4.0
asyncio~=3.4.3
beautifulsoup4~=4.12.3
bs4~=0.0.2
openpyxl~=3.1.5
pandas~=2.2.2
pathlib2~=2.3.7.post1
python-dotenv~=1.0.1
requests~=2.32.3
sqlalchemy~=1.4.25